power: avoid implicit casts of int64_t to double
As suggested by clang-tidy, avoid implicit of int64_t to double.

BUG=b:258053283
TEST=Local Tests + CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4022520
Commit-Queue: Alan Green <avg@chromium.org>
Tested-by: Alan Green <avg@chromium.org>

+++ power_manager/powerd/system/thermal/cooling_device.cc
@@ -109,9 +109,12 @@ bool CoolingDevice::InitSysfsFile() {
               << ") cooling device: " << device_path_;
     type_ = ThermalDeviceType::kOtherCooling;
   }
-  threshold_fair_ = ceil(max_state * kScale.at(type_).fair);
-  threshold_serious_ = ceil(max_state * kScale.at(type_).serious);
-  threshold_critical_ = ceil(max_state * kScale.at(type_).critical);
+  threshold_fair_ =
+      ceil(static_cast<double>(max_state) * kScale.at(type_).fair);
+  threshold_serious_ =
+      ceil(static_cast<double>(max_state) * kScale.at(type_).serious);
+  threshold_critical_ =
+      ceil(static_cast<double>(max_state) * kScale.at(type_).critical);
 
   polling_file_.Init(polling_path_);
   return true;
